% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_get.R
\name{eml_get}
\alias{eml_get}
\title{eml_get}
\usage{
eml_get(x, element, ...)
}
\arguments{
\item{x}{an EML object or child/descendant object}

\item{element}{name of the element to be extracted.
If multiple occurrences are found, will extract all}

\item{...}{additional arguments}
}
\description{
eml_get
}
\examples{
\donttest{
f <- system.file("xsd/test", "eml-datasetWithUnits.xml", package = "EML")
eml <- read_eml(f)
eml_get(eml, "physical")
eml_get(eml, "attributeList")

## The first argument need not be an "eml" class, it could be a child element; e.g.
eml_get(eml$dataset$dataTable, "physical")
}
}
