% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_eml_units.R
\name{get_plural}
\alias{get_plural}
\title{Get Plural Unit}
\usage{
get_plural(unit, udunits_units = load_udunits()$udunits_units)
}
\arguments{
\item{unit}{(character) unit}

\item{udunits_units}{(data.frame) load_udunits()}
}
\value{
(character) plural form of unit
}
\description{
Gets the plural form of a unit
Code is based on ut_form_plural function in xml.c of udunits2
}
