% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_eml_units.R
\name{get_split_unit}
\alias{get_split_unit}
\title{Split Unit}
\usage{
get_split_unit(unit, all_units = mem_load_all_units())
}
\arguments{
\item{unit}{(character) unit}

\item{all_units}{(data.frame)}
}
\value{
(character) unit_split
}
\description{
Performs formating on unit and splits into components
}
