% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_software.R
\name{set_software}
\alias{set_software}
\title{set_software}
\usage{
set_software(codemeta)
}
\arguments{
\item{codemeta}{codemeta object, see examples}
}
\value{
an eml software element
}
\description{
set_software
}
\examples{
options(emld_db = "eml-2.1.1",
        emld_schemaLocation="eml://ecoinformatics.org/eml-2.1.1/")
cm <- jsonlite::read_json(system.file("extdata/codemeta.json", package="eml2"))
software <- set_software(cm)
my_eml <- eml$eml(packageId = "eml-1.2",system = "knb", software = software)
eml_validate(my_eml)

# write_eml(my_eml, "eml.xml")
}
