% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_constructor.R
\name{template_constructor}
\alias{template_constructor}
\title{eml_creator <- template_constructor("creator")}
\usage{
template_constructor(object)
}
\description{
eml_creator <- template_constructor("creator")
}
