% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_eml_units.R
\name{try_reformat_unit}
\alias{try_reformat_unit}
\title{Reformat Units}
\usage{
try_reformat_unit(unit, exponents, exponents_numeric,
  all_units = load_all_units())
}
\arguments{
\item{unit}{(character)}

\item{exponents}{(character)}

\item{exponents_numeric}{(character)}

\item{all_units}{(data.frame)}
}
\value{
(character) unit. Fails if cannot reformat
}
\description{
Will attempt to reformat unit utilizing units::deparse_unit(units::as.units(unit)).
Used to decifier units entered symbolically. Will return the long form of the unit.
}
\examples{
unit <- "kg^2*s/(km^3*mole)"
exponents <- c("square", "cubic")
exponents_numeric <- c("2", "3")
try_reformat_unit(unit, exponents, exponents_numeric)
}
