% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eml.R
\name{read_eml}
\alias{read_eml}
\title{read_eml}
\usage{
read_eml(x)
}
\arguments{
\item{x}{path to an EML file}
}
\value{
an emld object (list / S3 object)
}
\description{
read_eml
}
\examples{
f <- system.file("extdata", "example.xml", package = "emld")
eml <- read_eml(f)
}
